Add tests for EReference name change mono and multi valued
