+++ tests/org.eclipse.sirius.emfjson.tests/src/main/resources/unit/references/extendedmetadata/TestChangeReferenceNameMono.json
@@ -1,0 +1,27 @@
+{
+  "json": {
+    "version": "1.0",
+    "encoding": "utf-8"
+  },
+  "ns": {
+    "nodes": "http://www.obeo.fr/EMFJson"
+  },
+  "schemaLocation": {
+    "http://www.obeo.fr/EMFJson": "../../../nodes.ecore"
+  },
+  "content": [
+    {
+      "eClass": "nodes:NodeSingleValueReference",
+      "data": {
+        "name": "Node1",
+        "singleValuedReferenceOld": "Node2"
+      }
+    },
+    {
+      "eClass": "nodes:NodeSingleValueReference",
+      "data": {
+        "name": "Node2"
+      }
+    }
+  ]
+}
