+++ tests/org.eclipse.sirius.emfjson.tests/src/main/resources/unit/references/extendedmetadata/TestChangeReferenceNameMono.xmi
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="ASCII"?>
+<xmi:XMI xmi:version="2.0"
+    xmlns:xmi="http://www.omg.org/XMI"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns:nodes="http://www.obeo.fr/EMFJson"
+    xsi:schemaLocation="http://www.obeo.fr/EMFJson ../../../nodes.ecore">
+  <nodes:NodeSingleValueReference
+      name="Node1"
+      singleValuedReference="Node2"/>
+  <nodes:NodeSingleValueReference
+      name="Node2"/>
+</xmi:XMI>
