+++ tests/org.eclipse.sirius.emfjson.tests/src/main/resources/unit/references/extendedmetadata/TestChangeReferenceNameMulti.json
@@ -1,0 +1,36 @@
+{
+  "json": {
+    "version": "1.0",
+    "encoding": "utf-8"
+  },
+  "ns": {
+    "nodes": "http://www.obeo.fr/EMFJson"
+  },
+  "schemaLocation": {
+    "http://www.obeo.fr/EMFJson": "../../../nodes.ecore"
+  },
+  "content": [
+    {
+      "eClass": "nodes:NodeMultiValuedReference",
+      "data": {
+        "name": "Node1",
+        "multiValuedReference": [
+          "Node2",
+          "Node3"
+         ]
+      }
+    },
+    {
+      "eClass": "nodes:NodeMultiValuedReference",
+      "data": {
+        "name": "Node2"
+      }
+    },
+    {
+      "eClass": "nodes:NodeMultiValuedReference",
+      "data": {
+        "name": "Node3"
+      }
+    }
+  ]
+}
