+++ tests/org.eclipse.sirius.emfjson.tests/src/main/resources/unit/references/extendedmetadata/TestChangeReferenceNameMulti.xmi
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="ASCII"?>
+<xmi:XMI xmi:version="2.0"
+    xmlns:xmi="http://www.omg.org/XMI"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns:nodes="http://www.obeo.fr/EMFJson"
+    xsi:schemaLocation="http://www.obeo.fr/EMFJson ../../../nodes.ecore">
+  <nodes:NodeMultiValuedReference
+      name="Node1"
+      multiValuedReference="Node2 Node3"/>
+  <nodes:NodeMultiValuedReference
+      name="Node2"/>
+  <nodes:NodeMultiValuedReference
+      name="Node3"/>
+</xmi:XMI>
